Updated frontend, new images library and rounds sharing system

+++ Initiativetracker/app/src/main/java/com/dmc/initiativetracker/data/local/entity/CharacterEntity.kt
@@ -14,7 +14,7 @@ data class CharacterEntity(
 
     val playerName: String,
     val characterName: String,
-    val initiative: Int,
+    val initiative: Double,
 
     val imageUri: String?,
